move parser tests to its own file
